Added initial testing for rendering components, moved components into specific files
