+++ src/components/Sidebar/Navbar/Navbar.jsx
@@ -1,0 +1,7 @@
+export default function Navbar() {
+  return (
+    <nav className="backgroundCard h-16 my-2 text-white p-3" data-testid="home">
+      Home
+    </nav>
+  );
+}
