+++ src/components/Sidebar/UserPlaylists/UserPlaylists.jsx
@@ -1,0 +1,10 @@
+export default function UserPlaylists() {
+  return (
+    <div
+      className="backgroundCard h-16 my-2 text-white p-3"
+      data-testid="userPlaylist"
+    >
+      User playlist
+    </div>
+  );
+}
