+++ src/components/Sidebar/UserProfile/UserProfile.jsx
@@ -1,0 +1,10 @@
+export default function UserProfile() {
+  return (
+    <div
+      className="backgroundCard h-16 my-2 justify-self-start text-white p-3"
+      data-testid="userProfile"
+    >
+      User Profile
+    </div>
+  );
+}
